+++ servers/mcp-neo4j-cloud-aura-api/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 ### Changed
+* Security upgrade python from 3.11-slim to 3.13.8-slim
 
 ### Added